requirements.txt

google-api-core==2.23.0
google-auth==2.36.0
google-cloud-speech==2.28.1
googleapis-common-protos==1.66.0
pytube==15.0.0
pytubefix==8.2.0
SpeechRecognition==3.11.0
youtube-dl==2021.12.17
yt-dlp @ https://github.com/yt-dlp/yt-dlp/archive/master.tar.gz

